add data layer, models, network, api service, api factory

+++ app/src/main/java/com/art3mvp/newsclient/domain/StatisticItem.kt
@@ -6,7 +6,7 @@ import kotlinx.parcelize.Parcelize
 @Parcelize
 data class StatisticItem(
     val type: StatisticType,
-    val count:Int = 0
+    val count: Long = 0
 ): Parcelable
 
 enum class StatisticType {
